+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version 2.6.3 (in Entwicklung)
+- Neu: Einstellungen neu in einem Dialog. 
 - Neu: Reiter Bewertung ersetzt durch Dialog. 
 - Neu: Reiter Ausdruck ersetzt durch Dialog. 
 - Neu: Reiter Etikett ersetzt durch Dialog.